requirement.txt
feedparser
trafilatura>=1.6.3
beautifulsoup4
PyYAML
python-dateutil
requests>=2.31
lxml
charset-normalizer==3.3.2
beautifulsoup4